# expression: implement vectorized evaluation for `builtinExportSet*ArgSig`