--- a/expression/builtin_string_vec.go
+++ b/expression/builtin_string_vec.go
@@ -607,11 +607,48 @@ func (b *builtinUnHexSig) vecEvalString(input *chunk.Chunk, result *chunk.Column
 }
 
 func (b *builtinExportSet3ArgSig) vectorized() bool {
-	return false
+	return true
 }
 
+// evalString evals EXPORT_SET(bits,on,off).
+// See https://dev.mysql.com/doc/refman/5.7/en/string-functions.html#function_export-set
 func (b *builtinExportSet3ArgSig) vecEvalString(input *chunk.Chunk, result *chunk.Column) error {
-	return errors.Errorf("not implemented")
+	n := input.NumRows()
+	bits, err := b.bufAllocator.get(types.ETInt, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(bits)
+	if err := b.args[0].VecEvalInt(b.ctx, input, bits); err != nil {
+		return err
+	}
+	on, err := b.bufAllocator.get(types.ETString, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(on)
+	if err := b.args[1].VecEvalString(b.ctx, input, on); err != nil {
+		return err
+	}
+	off, err := b.bufAllocator.get(types.ETString, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(off)
+	if err := b.args[2].VecEvalString(b.ctx, input, off); err != nil {
+		return err
+	}
+	result.ReserveString(n)
+	i64s := bits.Int64s()
+	for i := 0; i < n; i++ {
+		if bits.IsNull(i) || on.IsNull(i) || off.IsNull(i) {
+			result.AppendNull()
+			continue
+		}
+		result.AppendString(exportSet(i64s[i], on.GetString(i), off.GetString(i),
+			",", 64))
+	}
+	return nil
 }
 
 func (b *builtinASCIISig) vectorized() bool {
@@ -846,11 +883,56 @@ func (b *builtinLocateBinary3ArgsSig) vecEvalInt(input *chunk.Chunk, result *chu
 }
 
 func (b *builtinExportSet4ArgSig) vectorized() bool {
-	return false
+	return true
 }
 
+// evalString evals EXPORT_SET(bits,on,off,separator).
+// See https://dev.mysql.com/doc/refman/5.7/en/string-functions.html#function_export-set
 func (b *builtinExportSet4ArgSig) vecEvalString(input *chunk.Chunk, result *chunk.Column) error {
-	return errors.Errorf("not implemented")
+	n := input.NumRows()
+	bits, err := b.bufAllocator.get(types.ETInt, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(bits)
+	if err := b.args[0].VecEvalInt(b.ctx, input, bits); err != nil {
+		return err
+	}
+	on, err := b.bufAllocator.get(types.ETString, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(on)
+	if err := b.args[1].VecEvalString(b.ctx, input, on); err != nil {
+		return err
+	}
+	off, err := b.bufAllocator.get(types.ETString, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(off)
+	if err := b.args[2].VecEvalString(b.ctx, input, off); err != nil {
+		return err
+	}
+	separator, err := b.bufAllocator.get(types.ETString, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(separator)
+	if err := b.args[3].VecEvalString(b.ctx, input, separator); err != nil {
+		return err
+	}
+	result.ReserveString(n)
+	i64s := bits.Int64s()
+	for i := 0; i < n; i++ {
+		if bits.IsNull(i) || on.IsNull(i) || off.IsNull(i) || separator.IsNull(i) {
+			result.AppendNull()
+			continue
+		}
+		result.AppendString(exportSet(i64s[i], on.GetString(i), off.GetString(i),
+			separator.GetString(i), 64))
+	}
+	return nil
 }
 
 func (b *builtinRpadBinarySig) vectorized() bool {
@@ -1119,11 +1201,66 @@ func (b *builtinInsertSig) vecEvalString(input *chunk.Chunk, result *chunk.Colum
 }
 
 func (b *builtinExportSet5ArgSig) vectorized() bool {
-	return false
+	return true
 }
 
+// evalString evals EXPORT_SET(bits,on,off,separator,number_of_bits).
+// See https://dev.mysql.com/doc/refman/5.7/en/string-functions.html#function_export-set
 func (b *builtinExportSet5ArgSig) vecEvalString(input *chunk.Chunk, result *chunk.Column) error {
-	return errors.Errorf("not implemented")
+	n := input.NumRows()
+	bits, err := b.bufAllocator.get(types.ETInt, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(bits)
+	if err := b.args[0].VecEvalInt(b.ctx, input, bits); err != nil {
+		return err
+	}
+	on, err := b.bufAllocator.get(types.ETString, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(on)
+	if err := b.args[1].VecEvalString(b.ctx, input, on); err != nil {
+		return err
+	}
+	off, err := b.bufAllocator.get(types.ETString, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(off)
+	if err := b.args[2].VecEvalString(b.ctx, input, off); err != nil {
+		return err
+	}
+	separator, err := b.bufAllocator.get(types.ETString, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(separator)
+	if err := b.args[3].VecEvalString(b.ctx, input, separator); err != nil {
+		return err
+	}
+	numberOfBits, err := b.bufAllocator.get(types.ETInt, n)
+	if err != nil {
+		return err
+	}
+	defer b.bufAllocator.put(numberOfBits)
+	if err := b.args[4].VecEvalString(b.ctx, input, numberOfBits); err != nil {
+		return err
+	}
+	result.ReserveString(n)
+	bits.MergeNulls(numberOfBits)
+	i64s := bits.Int64s()
+	i64s2 := numberOfBits.Int64s()
+	for i := 0; i < n; i++ {
+		if bits.IsNull(i) || on.IsNull(i) || off.IsNull(i) || separator.IsNull(i) {
+			result.AppendNull()
+			continue
+		}
+		result.AppendString(exportSet(i64s[i], on.GetString(i), off.GetString(i),
+			separator.GetString(i), i64s2[i]))
+	}
+	return nil
 }
 
 func (b *builtinSubstring3ArgsSig) vectorized() bool {