PCP #12106
<!--
Thank you for contributing to TiDB! Please read TiDB's [CONTRIBUTING](https://github.com/pingcap/community/blob/master/CONTRIBUTING.md) document **BEFORE** filing this PR.
-->

### What problem does this PR solve? <!--add issue link with summary if exists-->
implement vectorized evaluation for `builtinExportSet*ArgSig`, for #12106

### What is changed and how it works?
```
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinStringFunc/builtinExportSet3ArgSig-VecBuiltinFunc-3                100          11189283 ns/op         9208039 B/op     129024 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinExportSet3ArgSig-NonVecBuiltinFunc-3              99          11763305 ns/op         9207929 B/op     129024 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinExportSet4ArgSig-VecBuiltinFunc-3                100          10818161 ns/op         9208019 B/op     129024 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinExportSet4ArgSig-NonVecBuiltinFunc-3             100          10978643 ns/op         9207918 B/op     129024 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinExportSet5ArgSig-VecBuiltinFunc-3                967           1246358 ns/op          262147 B/op      18432 allocs/op
BenchmarkVectorizedBuiltinStringFunc/builtinExportSet5ArgSig-NonVecBuiltinFunc-3             866           1344990 ns/op          262147 B/op      18432 allocs/op
PASS
ok      github.com/pingcap/tidb/expression      7.234s
```

### Check List <!--REMOVE the items that are not applicable-->

Tests <!-- At least one of them must be included. -->

- Unit test